DMP-4759 Improve AssociatedObjectDataExpiryDeletionAutomatedTask
Moved code into new service class so that @Transactional is effective
