+++ src/main/java/uk/gov/hmcts/darts/datamanagement/service/AssociatedObjectDataExpiryDeleterService.java
@@ -1,0 +1,7 @@
+package uk.gov.hmcts.darts.datamanagement.service;
+
+public interface AssociatedObjectDataExpiryDeleterService {
+
+    void delete(Integer batchSize);
+
+}
